refactor: Final release
