+++ src/type/timetable.ts
